refactor: changed location input view to bottomsheet
