+++ app/src/main/res/drawable/bottomsheet_handle.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="#88000000"/>
+    <corners android:radius="30dp"/>
+</shape>
