+++ app/src/main/res/drawable/round_bottomsheet.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="#efefef"/>
+    <corners android:topLeftRadius="30dp"
+        android:topRightRadius="30dp"/>
+</shape>
