+++ app/src/main/res/layout/fragment_kakao_map.xml
@@ -10,7 +10,7 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toTopOf="@id/text_location_name"
+        app:layout_constraintBottom_toBottomOf="parent"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:id="@+id/kakao_map_view"/>
@@ -50,30 +50,6 @@
         />
 
 
-    <TextView
-        android:textSize="24dp"
-        android:gravity="bottom|start"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintBottom_toTopOf="@id/text_location_address"
-        app:layout_constraintTop_toBottomOf="@id/kakao_map_view"
-        android:id="@+id/text_location_name"
-        android:paddingStart="15dp"
-        android:paddingBottom="5dp"
-        android:layout_width="0dp"
-        android:layout_height="60dp"/>
-    
-    <TextView
-        android:textSize="18dp"
-        android:gravity="top|start"
-        android:paddingStart="15dp"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/text_location_name"
-        android:id="@+id/text_location_address"
 
-        android:layout_width="0dp"
-        android:layout_height="40dp"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
